Flytter avbrytelse og utløp av oppgave til OppgaveService.
